% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.crd.R
\name{anova.crd}
\alias{anova.crd}
\title{One factor Completely Randomized Design}
\usage{
\method{anova}{crd}(
  treat,
  resp,
  quali = TRUE,
  graph = TRUE,
  graph.type = "barplotb",
  graph.col = "skyblue",
  xlab = "Treatments",
  ylab = "Variable",
  main = "",
  graph.den = 2,
  mcomp = "tukey",
  best.fit = TRUE,
  sigF = 0.05
)
}
\arguments{
\item{treat}{Numeric or complex vector containing the
treatments.}

\item{resp}{Numeric or complex vector containing the
response variable.}

\item{quali}{Logic. If TRUE (default), the treatments are
assumed qualitative, if FALSE, quantitatives.}

\item{mcomp}{Allows choosing the multiple comparison test;
the \emph{default} is the test of Tukey, however, the
options are: the LSD test ('lsd'), the test of Duncan
('duncan'), the test of Student-Newman-Keuls ('snk'),
the test of Scott-Knot ('scottkenott').}

\item{sigF}{The signficance to be used for the F test of
ANOVA; the default is 5\%.}
}
\value{
The output contains the ANOVA of the CRD, the
Shapiro-Wilk normality test for the residuals of the model,
the fitted regression models (when the treatments are
quantitative) and/or the multiple comparison tests (when
the treatments are qualitative).
}
\description{
\code{crd} Analyses balanced experiments in Completely
Randomized Design under one single factor, considering a
fixed model.
}
\examples{
dados=read.table("R/data.crd.csv", sep=";", header=T, dec=",")
attach(dados)
anova.crd(trat, resp)
}
\author{
Cícero Carlos de Souza Almeida
}
