% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.rbd.R
\name{anova.rbd}
\alias{anova.rbd}
\title{One factor Completely block Design}
\usage{
\method{anova}{rbd}(
  treat,
  resp,
  block,
  quali = TRUE,
  graph = TRUE,
  graph.type = "barplotb",
  graph.col = "skyblue",
  xlab = "Treatments",
  ylab = "Variable",
  main = "",
  graph.den = 2,
  mcomp = "tukey",
  best.fit = TRUE,
  sigF = 0.05
)
}
\arguments{
\item{treat}{Numeric or complex vector containing the
treatments.}

\item{resp}{Numeric or complex vector containing the
response variable.}

\item{block}{numeric or complex vector containing the block.}

\item{quali}{Logic. If TRUE (default), the treatments are
assumed qualitative, if FALSE, quantitatives.}

\item{mcomp}{Allows choosing the multiple comparison test;
the \emph{default} is the test of Tukey, however, the
options are: the LSD test ('lsd'), the LSD test with
Bonferroni protection ('lsdb'), the test of Duncan
('duncan'), the test of Student-Newman-Keuls ('snk'),
the test of Scott-Knot ('scottkenott').}

\item{sigF}{The signficance to be used for the F test of
ANOVA; the default is 5\%.}
}
\value{
The output contains the ANOVA of the CRD, the
Shapiro-Wilk normality test for the residuals of the model,
the fitted regression models (when the treatments are
quantitative) and/or the multiple comparison tests (when
the treatments are qualitative).
FERREIRA, E. B.; CAVALCANTI, P. P. Funcao em codigo R para
analisar experimentos em DIC simples, em uma so rodada. In:
REUNIAO ANUAL DA REGIAO BRASILEIRA DA SOCIEDADE
INTERNACIONAL DE BIOMETRIA, 54./SIMPOSIO DE ESTATISTICA
APLICADA A EXPERIMENTACAO AGRONOMICA, 13., 2009, Sao Carlos.
Programas e resumos... Sao Carlos, SP: UFSCar, 2009. p. 1-5.
}
\description{
\code{crd} Analyses balanced experiments in Completely
Randomized Design under one single factor, considering a
fixed model.
}
\examples{
dados=read.table("R/data.rbd.csv", sep=";", header=T, dec=",")
attach(dados)
anova.rbd(trat,resp,block)
}
\seealso{
\code{\link{fat2.crd}}, \code{\link{fat3.crd}},
\code{\link{split2.crd}}, \code{\link{fat2.ad.crd}} and
\code{\link{fat3.ad.crd}}.
}
\author{
Eric B Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}

Denismar Alves Nogueira

Portya Piscitelli Cavalcanti

Cícero Carlos de Souza Almeida
}
