% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.sp2.crd.R
\name{anova.sp2.crd}
\alias{anova.sp2.crd}
\title{Split-plots in CRD}
\usage{
\method{anova}{sp2.crd}(
  factor1,
  factor2,
  repet,
  resp,
  quali = c(TRUE, TRUE),
  fac.fix = "fac.A",
  fac.names = c("fac.A", "fac.B"),
  graph = TRUE,
  graph.type = "barplotb",
  graph.col = "skyblue",
  xlab = "Treatments",
  ylab = "Variable",
  graph.den = 2,
  mcomp = "tukey",
  sigF = 0.05
)
}
\arguments{
\item{factor1}{Numeric or complex vector containing the
factor 1 levels.}

\item{factor2}{Numeric or complex vector containing the
factor 2 levels.}

\item{repet}{Numeric or complex vector containing the
replications.}

\item{resp}{Numeric or complex vector containing the
response variable.}

\item{quali}{Logic. If TRUE (default), the treatments
are assumed qualitative, if FALSE, quantitatives.}

\item{fac.names}{Allows labeling the factors 1 and 2.}

\item{mcomp}{Allows choosing the multiple comparison
test; the \emph{default} is the test of Tukey, however,
the options are: the LSD test ('lsd'), the LSD test
with Bonferroni protection ('lsdb'), the test of Duncan
('duncan'), the test of Student-Newman-Keuls ('snk'),
the test of Scott-Knott ('sk'), the Calinski and
Corsten test ('ccF') and bootstrap multiple comparison's
test ('ccboot').}

\item{sigF}{The signficance to be used for the F test
of ANOVA; the default is 5\%.}

\item{sigT}{The signficance to be used for the multiple
comparison test; the default is 5\%.}
}
\value{
The output contains the ANOVA of the referred
CRD, the Shapiro-Wilk normality test for the residuals
of the model, the fitted regression models (when the
treatments are quantitative) and/or the multiple
comparison tests (when the treatments are qualitative).
}
\description{
\code{split2.crd} Analyses experiments in Split-plot scheme
in balanced Completely Randomized Design, considering a
fixed model.
}
\details{
The arguments sigT and mcomp will be used only
when the treatment are qualitative.
}
\note{
The \code{\link{graphics}} can be used to
construct regression plots and \code{\link{plotres}}
for residuals plots.
}
\examples{
data(ex9)
attach(ex9)
split2.crd(cobertura, prof, rep, pH, quali = c(TRUE, TRUE),
mcomp = "lsd", fac.names = c("Cover", "Depth"), sigT = 0.05,
sigF = 0.05)
}
\references{
BANZATTO, D. A.; KRONKA, S. N.
Experimentacao Agricola. 4 ed. Jaboticabal: Funep.
2006. 237 p.
}
\seealso{
\code{\link{split2.rbd}} and \code{\link{strip}}.
}
\author{
Eric B Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}

Denismar Alves Nogueira

Portya Piscitelli Cavalcanti
}
