% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ari_mean.R
\name{ari_mean}
\alias{ari_mean}
\title{Calcula a média aritmétrica para um conjunto de dados}
\usage{
ari_mean(x, na.rm = TRUE)
}
\value{
Valor da média
}
\description{
Função para calcular a média aritmétrica
}
\details{
Calcular a média geométrica para valores positivos
}
\examples{
data=c(3,4,4,4,5,5,6,8,8,9,9,10,11,12)
ari_mean(data)
}
\author{
Cicero Almeida
}
