% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des.crd.R
\name{des.crd}
\alias{des.crd}
\title{Densing for One Facgtor Completely Randomized Design}
\usage{
des.crd(
  trt,
  rep,
  col.1 = "gray",
  col.2 = "white",
  graph.col = TRUE,
  save = FALSE
)
}
\arguments{
\item{trt}{Numeric containing the numbers of treataments}

\item{rep}{Numeric containing the numbers of replications}

\item{col.1}{Allows choosing the color 1}

\item{col.2}{Allows choosing the color 2}

\item{graph.col}{Logic. If TRUE (default), plot the graph}

\item{save}{Logic. If TRUE (FALSE is default), save the Desing using the file names: "des.crd.tiff"}
}
\description{
Densing for One Facgtor Completely Randomized Design
}
\examples{
des.crd(trt=4, rep=3)
}
