% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{desc}
\alias{desc}
\title{Descriptive statistics}
\usage{
desc(
  y,
  x = 1,
  boxplot = FALSE,
  xlab = "Treatements",
  ylab = "Variable",
  graph.col = "skyblue"
)
}
\value{
Descriptive statistics
}
\description{
function for calculation of the Descriptive statistics
}
\examples{
data=c(3,4,4,4,5,5,6,8,8,9,9,10,11,12)
desc(data)
}
\author{
Cicero Almeida
}
