% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqtable.R
\name{freqtable}
\alias{freqtable}
\title{frequency table}
\usage{
freqtable(y, bins)
}
\value{
frequency table
}
\description{
function for frequency table
}
\examples{
data=c(3,4,4,4,5,5,6,8,8,9,9,10,11,12)
freqtable(data, bins)
}
\author{
Cicero Almeida
}
