% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_mean.R
\name{geo_mean}
\alias{geo_mean}
\title{Calcula a média geométrica para um conjunto de dados (x > 0)}
\usage{
geo_mean(x, na.rm = TRUE)
}
\value{
Valor da média
}
\description{
Função para calcular a média geométrica
}
\details{
Calcular a média geométrica para valores positivos
}
\examples{
data=c(3,4,4,4,5,5,6,8,8,9,9,10,11,12)
geo_mean(data)
}
\author{
Cicero Almeida
}
