% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har_mean.R
\name{har_mean}
\alias{har_mean}
\title{Calcula a média harmôinca para um conjunto de dados (x > 0)}
\usage{
har_mean(x, na.rm = TRUE)
}
\value{
Valor da média
}
\description{
Função para calcular a média harmônica
}
\details{
Calcular a média geométrica para valores positivos
}
\examples{
data=c(3,4,4,4,5,5,6,8,8,9,9,10,11,12)
har_mean(data)
}
\author{
Cicero Almeida
}
