% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histg.R
\name{histg}
\alias{histg}
\title{Histogram with graph}
\usage{
histg(
  y,
  xlab = "Distribution",
  ylab = "Frequency",
  graph.col = "turquoise",
  scores = "empiric",
  border = TRUE,
  freq = TRUE,
  graph.type = "base",
  main = ""
)
}
\value{
frequency table and histogram graph
}
\description{
function for calculation of table freq and histogram
}
\examples{
data=c(3,4,4,4,5,5,6,8,8,9,9,10,11,12)
histg(data)
}
\author{
Cicero Almeida
}
