% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pon_mean.R
\name{pon_mean}
\alias{pon_mean}
\title{Calcula a média ponderada para um conjunto de dados, com os devidos peso}
\usage{
pon_mean(x, w, na.rm = TRUE)
}
\value{
Valor da média
}
\description{
Função para calcular a média ponderada
}
\details{
Calcular a média ponderada
}
\examples{
data=c(3,4,4,4,5,5,6,8,8,9,9,10,11,12)
w=c(5,7,6,5,2,1,5,4,8,3,2,3,8,9)
pon_mean(data,w)
}
\author{
Cicero Almeida
}
